[SL] Refactor api calls into a services directory
